pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polybot"
version = "0.1.0"
description = "Polymarket bot detection and copy-trading system"
requires-python = ">=3.10"
dependencies = [
    "py-clob-client>=0.34",
    "websocket-client>=1.6",
    "requests>=2.31",
    "click>=8.1",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "aiosqlite>=0.19",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "pytest-mock>=3.11",
]

[project.scripts]
polybot = "polybot.cli:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
include = ["polybot*"]
